% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab8shiny.R
\name{lab8shiny}
\alias{lab8shiny}
\title{lab8shiny}
\usage{
lab8shiny()
}
\value{
A web-browser app to simulate the density of the objective function compared to the theoretical where the number of iterations are user-chosen
}
\description{
A shiny app which takes a chosen number of iterations, and simulates an objective function B, where B~Gamma(2,1)
}
\examples{
\dontrun{lab8shiny()}
}
