% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysimA.R
\name{mysimA}
\alias{mysimA}
\title{mysimA}
\usage{
mysimA(iter)
}
\arguments{
\item{iter}{Number of iterations for the simulation}
}
\value{
a histogram of the simulation and a named list containing: quantiles, mean, variance (of the sim), and the theoretical variance.
}
\description{
This function simulations a Uniform(0,1) Distribution
}
\examples{
\dontrun{mysimA(iter=10000)}
}
