% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysimB.R
\name{mysimB}
\alias{mysimB}
\title{mysimB}
\usage{
mysimB(iter)
}
\arguments{
\item{iter}{Number of iterations for the simulation}
}
\value{
a histogram of the simulation with a theoretical curve and a named list containing: quantiles, mean, variance (of the sim), and the theoretical variance.
}
\description{
This function simulations a Gamma(alpha=2, beta=1) Distribution
}
\examples{
\dontrun{mysimB(iter=10000)}
}
