% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmyF.R
\name{rmyF}
\alias{rmyF}
\title{rmyF}
\usage{
rmyF(iter)
}
\arguments{
\item{iter}{iter Number of iterations for the simulation}
}
\value{
a histogram of the simulation with a theoretical curve
}
\description{
a w-F theorem simulation of a Gamma(alpha=2, beta=1) using an objective function approach.
}
\examples{
\dontrun{rmyF(iter=10000)}
}
